+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -70,6 +70,7 @@ object Constants {
         const val INTELLIJ_PLATFORM_PLUGIN = "intellijPlatformPlugin"
         const val INTELLIJ_PLATFORM_BUNDLED_PLUGINS = "intellijPlatformBundledPlugins"
         const val INTELLIJ_PLATFORM_BUNDLED_MODULES = "intellijPlatformBundledModules"
+        const val INTELLIJ_PLATFORM_CLASSPATH = "intellijPlatformClasspath"
         const val INTELLIJ_PLATFORM_DEPENDENCIES = "intellijPlatformDependencies"
         const val INTELLIJ_PLATFORM_JAVA_COMPILER = "intellijPlatformJavaCompiler"
         const val INTELLIJ_PLATFORM_TEST_DEPENDENCIES = "intellijPlatformTestDependencies"